# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JSON-Mate - Документация API",
        "description": "Документация API для управления задачами",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить список всех задач",
                "description": "Получить все задачи",
                "operationId": "6a93d54730cde509d735e538ab6800f6",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск задач по названию",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_completed",
                        "in": "query",
                        "description": "Фильтр по выполненным задачам (0 - не выполнено, 1 - выполнено)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Фильтр по приоритету (низкий, средний, высокий)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Фильтр по категории",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создать новую задачу",
                "description": "Создать новую задачу",
                "operationId": "9b481c0ed5fdfa1a592d70ba39a348ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Task"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Созданная задача",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить задачу по ID",
                "description": "Получить одну задачу",
                "operationId": "fd3677b5b24695d7b801715869b1357c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные задачи",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Обновить задачу",
                "description": "Обновить задачу",
                "operationId": "85377ff0e935f093dd75ea6886923dbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Task"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновленная задача",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить задачу по ID",
                "description": "Удалить задачу",
                "operationId": "5ed5dfd4c645e9398f4282969bfe6478",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача удалена"
                    },
                    "404": {
                        "description": "Задача не найдена"
                    }
                }
            }
        },
        "/tasks/delete-multiple": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить несколько задач",
                "description": "Удалить несколько задач",
                "operationId": "838ba881df5d14e52cc8cdc1f694a8c6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Задачи удалены"
                    },
                    "422": {
                        "description": "Задачи не найдены"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "title": "Task",
                "description": "Модель задачи",
                "required": [
                    "title"
                ],
                "properties": {
                    "id": {
                        "description": "ID задачи",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Название задачи",
                        "type": "string",
                        "example": "Название задачи"
                    },
                    "description": {
                        "description": "Описание задачи",
                        "type": "string",
                        "example": "Описание задачи"
                    },
                    "due_date": {
                        "description": "Срок выполнения",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-15T10:00:00Z",
                        "nullable": true
                    },
                    "priority": {
                        "description": "Приоритет",
                        "type": "string",
                        "enum": [
                            "низкий",
                            "средний",
                            "высокий"
                        ],
                        "example": "средний"
                    },
                    "category": {
                        "description": "Категория задачи",
                        "type": "string",
                        "example": "Работа",
                        "nullable": true
                    },
                    "is_completed": {
                        "description": "Статус выполнения",
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Операции с задачами"
        }
    ]
}